+++ CHANGELOG.md
@@ -16,6 +16,7 @@ use [Keep a Changelog] (http://keepachangelog.com/).
 - `ArrayHelper::column()` with non-Closure `$index_key` was incorrectly assigning to `$finalValue` instead of `$finalIndex`
 - `FileHelper::uniformizePathSeparator()` was removing double slashes instead of collapsing them, corrupting paths
 - `b_array_simple()` bootstrap function was ignoring the `$prefix` parameter due to accidental reassignment to `null`
+- `ObjectHelper::getPropertyValue()` always set `$exists` to `false` after a successful `__call()` due to `finally` block
 
 ## [1.12.0] - 2025-05-27